requirements.txt
Django==3.1.2
djangorestframework==3.12.1
importlib-metadata==2.0.0
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==3.10.0
pytz==2020.1
Shapely==1.7.1
sqlparse==0.3.1
geos==3.8.1
proj==7.1.1
curl==7.72.0
gdal==3.1.3
postgis==3.0.3
tiff==4.1.0
nodejs==8.10
npm==3.5.2
leaflet~=0.0.3
django-leaflet~=0.27.1